requirements.txt
streamlit
pypdf2
python-docx